broke text.js into separate files for each item
